move around some Discord code, try to separate code into more distinct categories Also remove the locally stored collection of active channels, was never used, only updated
