+++ src/discord/permission.ts
@@ -13,7 +13,7 @@ export async function checkPermission(member: GuildMember, permission: Permissio
         case PermissionLevel.ADMINISTRATOR: {
             if (member.user.bot || member.permissions.has("Administrator")) return true
             const adminRole = needNotNullOrUndefined(await member.guild.roles.fetch(await needSetting("admin_role_snowflake")), "adminRole")
-            return (member.roles.cache.some((role, snowflake) => snowflake === adminRole.id))
+            return (member.roles.cache.some((_role, snowflake) => snowflake === adminRole.id))
         }
         case PermissionLevel.HUSANSVARLIG: {
             const husansvarligRole = needNotNullOrUndefined(await member.guild.roles.fetch(needEnvVariable(EnvironmentVariable.HUSANSVARLIG_ROLE_SNOWFLAKE)), "husansvarligRole")
